requirements.txt
pyyaml>=6.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
openai>=1.3.0
google-generativeai>=0.3.0
pillow>=10.0.0
opencv-python>=4.8.0
pandas>=2.0.0
jinja2>=3.1.0
weasyprint>=60.0
python-multipart>=0.0.6
aiofiles>=23.0.0
httpx>=0.25.0
